refactor(#32): 닉네임 검증을 사용자 부가기능으로 분리
