+++ src/main/java/deepdivers/community/domain/member/controller/open/AccountOpenController.java
@@ -8,9 +8,11 @@ import deepdivers.community.domain.member.service.AccountService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
@@ -19,6 +21,12 @@ import org.springframework.web.bind.annotation.RestController;
 public class AccountOpenController implements AccountOpenControllerDocs {
 
     private final AccountService accountService;
+
+    @GetMapping("/verify/nicknames")
+    public ResponseEntity<NoContent> verifyNickname(@RequestParam final String nickname) {
+        final NoContent response = accountService.verifyNickname(nickname);
+        return ResponseEntity.ok(response);
+    }
 
     @PostMapping("/verify/emails")
     public ResponseEntity<NoContent> verifyEmail(@RequestBody @Valid final VerifyEmailRequest request) {
