+++ src/test/java/deepdivers/community/domain/member/controller/MemberOpenControllerTest.java
@@ -12,6 +12,7 @@ import deepdivers.community.domain.ControllerTest;
 import deepdivers.community.domain.common.API;
 import deepdivers.community.domain.common.NoContent;
 import deepdivers.community.domain.member.controller.api.MemberApiController;
+import deepdivers.community.domain.member.controller.open.AccountOpenController;
 import deepdivers.community.domain.member.controller.open.MemberOpenOpenController;
 import deepdivers.community.domain.member.dto.request.MemberLoginRequest;
 import deepdivers.community.domain.member.dto.request.MemberProfileRequest;
@@ -20,6 +21,7 @@ import deepdivers.community.domain.member.dto.response.ImageUploadResponse;
 import deepdivers.community.domain.member.dto.response.MemberProfileResponse;
 import deepdivers.community.domain.member.dto.response.statustype.MemberStatusType;
 import deepdivers.community.domain.member.model.Member;
+import deepdivers.community.domain.member.service.AccountService;
 import deepdivers.community.domain.token.dto.TokenResponse;
 import io.restassured.common.mapper.TypeRef;
 import io.restassured.http.ContentType;
@@ -27,17 +29,23 @@ import io.restassured.module.mockmvc.RestAssuredMockMvc;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.web.context.WebApplicationContext;
 import org.springframework.web.multipart.MultipartFile;
 
 @WebMvcTest(controllers = {
-        MemberOpenOpenController.class,
-        MemberApiController.class
+    MemberOpenOpenController.class,
+    MemberApiController.class,
+    AccountOpenController.class
 })
 class MemberOpenControllerTest extends ControllerTest {
+
+    @MockBean
+    private AccountService accountService;
 
     @BeforeEach
     void setUp(WebApplicationContext webApplicationContext) throws Exception {
@@ -49,7 +57,8 @@ class MemberOpenControllerTest extends ControllerTest {
     @DisplayName("회원가입 요청이 성공적으로 처리되면 200 OK와 함께 응답을 반환한다")
     void signUpSuccessfullyReturns200OK() {
         // given
-        MemberSignUpRequest request = new MemberSignUpRequest("test@email.com", "test1234!", "test", "test", "010-1234-5678");
+        MemberSignUpRequest request = new MemberSignUpRequest("test@email.com", "test1234!", "test", "test",
+            "010-1234-5678");
 
         Member account = Member.of(request, this.encryptor);
         NoContent mockResponse = NoContent.from(MemberStatusType.MEMBER_SIGN_UP_SUCCESS);
@@ -57,14 +66,15 @@ class MemberOpenControllerTest extends ControllerTest {
 
         // when
         NoContent response = RestAssuredMockMvc
-                .given().log().all()
-                .contentType(MediaType.APPLICATION_JSON)
-                .body(request)
-                .when().post("/members/sign-up")
-                .then().log().all()
-                .status(HttpStatus.OK)
-                .extract()
-                .as(new TypeRef<>(){});
+            .given().log().all()
+            .contentType(MediaType.APPLICATION_JSON)
+            .body(request)
+            .when().post("/members/sign-up")
+            .then().log().all()
+            .status(HttpStatus.OK)
+            .extract()
+            .as(new TypeRef<>() {
+            });
 
         // then
         assertThat(response).isNotNull();
@@ -79,50 +89,52 @@ class MemberOpenControllerTest extends ControllerTest {
 
         // when, then
         RestAssuredMockMvc
-                .given().log().all()
-                .contentType(MediaType.APPLICATION_JSON)
-                .body(request)
-                .when().post("/members/sign-up")
-                .then().log().all()
-                .status(HttpStatus.BAD_REQUEST)
-                .body("code", equalTo(101))
-                .body("message", containsString("사용자 이메일 정보가 필요합니다."));
+            .given().log().all()
+            .contentType(MediaType.APPLICATION_JSON)
+            .body(request)
+            .when().post("/members/sign-up")
+            .then().log().all()
+            .status(HttpStatus.BAD_REQUEST)
+            .body("code", equalTo(101))
+            .body("message", containsString("사용자 이메일 정보가 필요합니다."));
     }
 
     @Test
     @DisplayName("회원가입 요청 시 닉네임 정보가 존재하지 않는다면 400 BadRequest 를 반환한다.")
     void signUpNullNicknameReturns400BadRequest() {
         // given
-        MemberSignUpRequest request = new MemberSignUpRequest("test@email.com", "test1234!", null, "테스트", "010-1234-5678");
+        MemberSignUpRequest request = new MemberSignUpRequest("test@email.com", "test1234!", null, "테스트",
+            "010-1234-5678");
 
         // when, then
         RestAssuredMockMvc
-                .given().log().all()
-                .contentType(MediaType.APPLICATION_JSON)
-                .body(request)
-                .when().post("/members/sign-up")
-                .then().log().all()
-                .status(HttpStatus.BAD_REQUEST)
-                .body("code", equalTo(101))
-                .body("message", containsString("사용자 닉네임 정보가 필요합니다."));
+            .given().log().all()
+            .contentType(MediaType.APPLICATION_JSON)
+            .body(request)
+            .when().post("/members/sign-up")
+            .then().log().all()
+            .status(HttpStatus.BAD_REQUEST)
+            .body("code", equalTo(101))
+            .body("message", containsString("사용자 닉네임 정보가 필요합니다."));
     }
 
     @Test
     @DisplayName("회원가입 요청 시 이미지 주소 정보가 존재하지 않는다면 400 BadRequest 를 반환한다.")
     void signUpNullImageUrlReturns400BadRequest() {
         // given
-        MemberSignUpRequest request = new MemberSignUpRequest("test@email.com", "test1234!", "테스트", null, "010-1234-5678");
+        MemberSignUpRequest request = new MemberSignUpRequest("test@email.com", "test1234!", "테스트", null,
+            "010-1234-5678");
 
         // when, then
         RestAssuredMockMvc
-                .given().log().all()
-                .contentType(MediaType.APPLICATION_JSON)
-                .body(request)
-                .when().post("/members/sign-up")
-                .then().log().all()
-                .status(HttpStatus.BAD_REQUEST)
-                .body("code", equalTo(101))
-                .body("message", containsString("사용자 이미지 정보가 필요합니다."));
+            .given().log().all()
+            .contentType(MediaType.APPLICATION_JSON)
+            .body(request)
+            .when().post("/members/sign-up")
+            .then().log().all()
+            .status(HttpStatus.BAD_REQUEST)
+            .body("code", equalTo(101))
+            .body("message", containsString("사용자 이미지 정보가 필요합니다."));
     }
 
     @Test
@@ -133,14 +145,14 @@ class MemberOpenControllerTest extends ControllerTest {
 
         // when, then
         RestAssuredMockMvc
-                .given().log().all()
-                .contentType(MediaType.APPLICATION_JSON)
-                .body(request)
-                .when().post("/members/sign-up")
-                .then().log().all()
-                .status(HttpStatus.BAD_REQUEST)
-                .body("code", equalTo(101))
-                .body("message", containsString("사용자 전화번호 정보가 필요합니다."));
+            .given().log().all()
+            .contentType(MediaType.APPLICATION_JSON)
+            .body(request)
+            .when().post("/members/sign-up")
+            .then().log().all()
+            .status(HttpStatus.BAD_REQUEST)
+            .body("code", equalTo(101))
+            .body("message", containsString("사용자 전화번호 정보가 필요합니다."));
     }
 
     @Test
@@ -155,14 +167,15 @@ class MemberOpenControllerTest extends ControllerTest {
 
         // when
         API<TokenResponse> response = RestAssuredMockMvc
-                .given().log().all()
-                .contentType(MediaType.APPLICATION_JSON)
-                .body(loginRequest)
-                .when().post("/members/login")
-                .then().log().all()
-                .status(HttpStatus.OK)
-                .extract()
-                .as(new TypeRef<>(){});
+            .given().log().all()
+            .contentType(MediaType.APPLICATION_JSON)
+            .body(loginRequest)
+            .when().post("/members/login")
+            .then().log().all()
+            .status(HttpStatus.OK)
+            .extract()
+            .as(new TypeRef<>() {
+            });
 
         // then
         assertThat(response).isNotNull();
@@ -173,21 +186,24 @@ class MemberOpenControllerTest extends ControllerTest {
     @DisplayName("프로필 조회 요청이 성공적으로 처리되면 200 OK와 함께 응답을 반환한다")
     void findProfileSuccessfullyReturns200OK() {
         // given
-        MemberSignUpRequest signUpRequest = new MemberSignUpRequest("test@email.com", "test1234!", "test", "test", "010-1234-5678");
+        MemberSignUpRequest signUpRequest = new MemberSignUpRequest("test@email.com", "test1234!", "test", "test",
+            "010-1234-5678");
         Member member = Member.of(signUpRequest, encryptor);
         MemberProfileResponse memberProfileResponse = MemberProfileResponse.from(member);
-        API<MemberProfileResponse> mockResponse = API.of(MemberStatusType.VIEW_OTHER_PROFILE_SUCCESS, memberProfileResponse);
+        API<MemberProfileResponse> mockResponse = API.of(MemberStatusType.VIEW_OTHER_PROFILE_SUCCESS,
+            memberProfileResponse);
         given(memberService.getProfile(any(Member.class), anyLong())).willReturn(mockResponse);
         Long profileOwnerId = 1L;
 
         // when
         API<MemberProfileResponse> response = RestAssuredMockMvc.given().log().all()
-                .pathParam("memberId", profileOwnerId)
-                .when().get("/api/members/{memberId}/me")
-                .then().log().all()
-                .status(HttpStatus.OK)
-                .extract()
-                .as(new TypeRef<>(){});
+            .pathParam("memberId", profileOwnerId)
+            .when().get("/api/members/{memberId}/me")
+            .then().log().all()
+            .status(HttpStatus.OK)
+            .extract()
+            .as(new TypeRef<>() {
+            });
 
         // then
         assertThat(response).isNotNull();
@@ -206,13 +222,14 @@ class MemberOpenControllerTest extends ControllerTest {
 
         // when
         API<ImageUploadResponse> response = RestAssuredMockMvc.given().log().all()
-                .contentType(MediaType.MULTIPART_FORM_DATA)
-                .multiPart("imageFile", contentBody, MediaType.IMAGE_PNG_VALUE)
-                .when().post("/api/members/profile-image")
-                .then().log().all()
-                .status(HttpStatus.OK)
-                .extract()
-                .as(new TypeRef<>(){});
+            .contentType(MediaType.MULTIPART_FORM_DATA)
+            .multiPart("imageFile", contentBody, MediaType.IMAGE_PNG_VALUE)
+            .when().post("/api/members/profile-image")
+            .then().log().all()
+            .status(HttpStatus.OK)
+            .extract()
+            .as(new TypeRef<>() {
+            });
 
         // then
         assertThat(response).isNotNull();
@@ -226,12 +243,12 @@ class MemberOpenControllerTest extends ControllerTest {
 
         // when, then
         RestAssuredMockMvc.given().log().all()
-                .contentType(MediaType.MULTIPART_FORM_DATA)
-                .when().post("/api/members/profile-image")
-                .then().log().all()
-                .status(HttpStatus.INTERNAL_SERVER_ERROR)
-                .body("code", equalTo(100))
-                .body("message", containsString("알 수 없는 서버 에러가 발생했습니다."));
+            .contentType(MediaType.MULTIPART_FORM_DATA)
+            .when().post("/api/members/profile-image")
+            .then().log().all()
+            .status(HttpStatus.INTERNAL_SERVER_ERROR)
+            .body("code", equalTo(100))
+            .body("message", containsString("알 수 없는 서버 에러가 발생했습니다."));
     }
 
     @Test
@@ -240,16 +257,17 @@ class MemberOpenControllerTest extends ControllerTest {
         // given
         String nickname = "안녕하세요";
         NoContent mockResponse = NoContent.from(MemberStatusType.NICKNAME_VALIDATE_SUCCESS);
-        given(memberService.verifyNickname(anyString())).willReturn(mockResponse);
+        given(accountService.verifyNickname(anyString())).willReturn(mockResponse);
 
         NoContent response = RestAssuredMockMvc.given().log().all()
             .contentType(ContentType.JSON)
             .queryParam("nickname", nickname)
-            .when().get("/members/verify/nicknames")
+            .when().get("/accounts/verify/nicknames")
             .then().log().all()
             .status(HttpStatus.OK)
             .extract()
-            .as(new TypeRef<>(){});
+            .as(new TypeRef<>() {
+            });
 
         // then
         assertThat(response).isNotNull();
@@ -263,7 +281,7 @@ class MemberOpenControllerTest extends ControllerTest {
         // when, then
         RestAssuredMockMvc.given().log().all()
             .contentType(ContentType.JSON)
-            .when().get("/members/verify/nicknames")
+            .when().get("/accounts/verify/nicknames")
             .then().log().all()
             .status(HttpStatus.BAD_REQUEST)
             .body("code", equalTo(203))
@@ -274,11 +292,13 @@ class MemberOpenControllerTest extends ControllerTest {
     @DisplayName("올바른 프로필 정보 수정시 200 OK가 떨어진다.")
     void updateProfileReturns200OK() {
         // given
-        MemberProfileRequest request = new MemberProfileRequest("test","test","","010-1234-5678","","");
-        MemberSignUpRequest signUpRequest = new MemberSignUpRequest("test@email.com", "test1234!", "test", "test", "010-1234-5678");
+        MemberProfileRequest request = new MemberProfileRequest("test", "test", "", "010-1234-5678", "", "");
+        MemberSignUpRequest signUpRequest = new MemberSignUpRequest("test@email.com", "test1234!", "test", "test",
+            "010-1234-5678");
         Member member = Member.of(signUpRequest, encryptor);
         MemberProfileResponse memberProfileResponse = MemberProfileResponse.from(member);
-        API<MemberProfileResponse> mockResponse = API.of(MemberStatusType.UPDATE_PROFILE_SUCCESS, memberProfileResponse);
+        API<MemberProfileResponse> mockResponse = API.of(MemberStatusType.UPDATE_PROFILE_SUCCESS,
+            memberProfileResponse);
         given(memberService.updateProfile(1L, request)).willReturn(mockResponse);
 
         // when
@@ -289,7 +309,8 @@ class MemberOpenControllerTest extends ControllerTest {
             .then().log().all()
             .status(HttpStatus.OK)
             .extract()
-            .as(new TypeRef<>(){});
+            .as(new TypeRef<>() {
+            });
 
         // then
         assertThat(response).isNotNull();
@@ -300,7 +321,7 @@ class MemberOpenControllerTest extends ControllerTest {
     @DisplayName("프로필 정보 수정시 닉네임 정보가 없다면 400 BadRequest 가 떨어진다.")
     void profileUpdateNullNicknameReturns400BadRequest() {
         // given
-        MemberProfileRequest request = new MemberProfileRequest("","test","","010-1234-5678","","");
+        MemberProfileRequest request = new MemberProfileRequest("", "test", "", "010-1234-5678", "", "");
 
         // when, then
         RestAssuredMockMvc.given().log().all()
@@ -317,7 +338,7 @@ class MemberOpenControllerTest extends ControllerTest {
     @DisplayName("프로필 정보 수정시 이미지 정보가 없다면 400 BadRequest 가 떨어진다.")
     void profileUpdateNullImageUrlReturns400BadRequest() {
         // given
-        MemberProfileRequest request = new MemberProfileRequest("test","","","010-1234-5678","","");
+        MemberProfileRequest request = new MemberProfileRequest("test", "", "", "010-1234-5678", "", "");
 
         // when, then
         RestAssuredMockMvc.given().log().all()
@@ -334,7 +355,7 @@ class MemberOpenControllerTest extends ControllerTest {
     @DisplayName("프로필 정보 수정시 전화번호 정보가 없다면 400 BadRequest 가 떨어진다.")
     void profileUpdateNullPhoneNumberReturns400BadRequest() {
         // given
-        MemberProfileRequest request = new MemberProfileRequest("test","test","","","","");
+        MemberProfileRequest request = new MemberProfileRequest("test", "test", "", "", "", "");
 
         // when, then
         RestAssuredMockMvc.given().log().all()
